ProteinChains

Documentation for ProteinChains.

ProteinChains.STANDARD_RESIDUE_TEMPLATE

ProteinChains.BackboneGeometry

ProteinChains.IdealResidue

ProteinChains.ProteinChain

ProteinChains.ProteinDataset

ProteinChains.ProteinStructure

ProteinChains.append_residue

ProteinChains.prepend_residue

ProteinChains.renumber!

ProteinChains.STANDARD_RESIDUE_TEMPLATE — Constant

STANDARD_RESIDUE_TEMPLATE

This is a template of a "standard residue", with a very specific and distinct shape, size, and orientation. which needs to be consistent if we want to represent protein structures as sets of residue rotations and translations.

Thus, we can use this residue as a template for aligning other residues with very precise geometry to it.

julia> IdealResidue{Float64}(BackboneGeometry(N_Ca_C_angle = 1.93); template=ProteinChains.STANDARD_RESIDUE_TEMPLATE)
3×3 IdealResidue{Float64}:
 -1.06447   -0.199174   1.26364
  0.646303  -0.529648  -0.116655
  0.0        0.0        0.0

ProteinChains.BackboneGeometry — Type

BackboneGeometry(;
    N_Ca_length = 1.46,
    Ca_C_length = 1.52,
    C_N_length = 1.33,

    N_Ca_C_angle = 1.94,
    Ca_C_N_angle = 2.03,
    C_N_Ca_angle = 2.13,
)

Define the idealized bond lengths and bond angles of a protein backbone.

ProteinChains.IdealResidue — Type

IdealResidue{T<:AbstractFloat} <: AbstractMatrix{T}

IdealResidue{T}(backbone_geometry=DEFAULT_BACKBONE_GEOMETRY; template=nothing) where T

A 3x3 matrix representing the idealized geometry of a protein residue, with columns representing the N, Ca, and C atom positions of a residue positioned at the origin.

ProteinChains.ProteinChain — Type

ProteinChain{T<:Real,Ps<:NamedProperties}

ProteinChains.ProteinDataset — Type

ProteinDataset <: AbstractDict{String,ProteinStructure}

ProteinChains.ProteinStructure — Type

ProteinStructure{T} <: AbstractVector{ProteinChain{T}}

ProteinChains.append_residue — Method

append_residue(Backbone::Backbone, torsion_angles::Vector{<:Real}; ideal::BackboneGeometry=DEFAULT_BACKBONE_GEOMETRY)

Create a new backbone by appending 3 new torsion angles (ψ, ω, ϕ) at the end, using bond lengths and bond angles specified in BackboneGeometry.

ProteinChains.prepend_residue — Method

append_residue(Backbone::Backbone, torsion_angles::Vector{<:Real}; ideal::BackboneGeometry=DEFAULT_BACKBONE_GEOMETRY)

Create a new backbone by prepending 3 new torsion angles (ψ, ω, ϕ) at the beginning, using bond lengths and bond angles specified in the BackboneGeometry.

Note

The torsion angle order is the same as it would be when appending. The order is not reversed.

ProteinChains.renumber! — Method

renumber!(structure::ProteinStructure, mmcif_dict::BioStructures.MMCIFDict)

Renumber the residues in a ProteinStructure object according to the numbering aligned to a reference sequence in the MMCIF file.