ProteinChains

Documentation for ProteinChains.

ProteinChains.STANDARD_RESIDUE_TEMPLATE

ProteinChains.BackboneGeometry

ProteinChains.IdealResidue

ProteinChains.ProteinChain

ProteinChains.ProteinStructure

ProteinChains.ProteinStructureStore

ProteinChains.ProteinStructureStore

ProteinChains.ProteinStructureStore

ProteinChains.append_residue

ProteinChains.deleteproperty

ProteinChains.deserialize

ProteinChains.mapmmcif

ProteinChains.pdbentry

ProteinChains.prepend_residue

ProteinChains.readattribute

ProteinChains.readproperty

ProteinChains.serialize

ProteinChains.writeattribute

ProteinChains.writeproperty

ProteinChains.STANDARD_RESIDUE_TEMPLATE — Constant

STANDARD_RESIDUE_TEMPLATE

This is a template of a "standard residue", with a very specific and distinct shape, size, and orientation. which needs to be consistent if we want to represent protein structures as sets of residue rotations and translations.

Thus, we can use this residue as a template for aligning other residues with very precise geometry to it.

julia> IdealResidue{Float64}(BackboneGeometry(N_Ca_C_angle = 1.93); template=ProteinChains.STANDARD_RESIDUE_TEMPLATE)
3×3 IdealResidue{Float64}:
 -1.06447   -0.199174   1.26364
  0.646303  -0.529648  -0.116655
  0.0        0.0        0.0

ProteinChains.BackboneGeometry — Type

BackboneGeometry(;
    N_Ca_length = 1.46,
    Ca_C_length = 1.52,
    C_N_length = 1.33,

    N_Ca_C_angle = 1.94,
    Ca_C_N_angle = 2.03,
    C_N_Ca_angle = 2.13,
)

Define the idealized bond lengths and bond angles of a protein backbone.

ProteinChains.IdealResidue — Type

IdealResidue{T<:AbstractFloat} <: AbstractMatrix{T}

IdealResidue{T}(backbone_geometry=DEFAULT_BACKBONE_GEOMETRY; template=nothing) where T

A 3x3 matrix representing the idealized geometry of a protein residue, with columns representing the N, Ca, and C atom positions of a residue positioned at the origin.

ProteinChains.ProteinChain — Type

ProteinChain{T<:Real}

Represents a protein chain with a basic set of fields from which some other properties might be derived.

Fields

id::String: Identifier for the protein chain.

atoms::Vector{Vector{Atom{T}}}: List of atoms in each residue.

sequence::String: Amino acid sequence of the protein.

numbering::Vector{Int}: Residue numbering (author).

ins_codes::String: Insertion codes for each residue.

```

ProteinChains.ProteinStructure — Type

ProteinStructure{T} <: AbstractVector{ProteinChain{T}}

Fields

name::String: Usually just the base name of the original file.

atoms::Vector{Atom{T}}: free atoms from the structure that were not part of any protein residue.

chains::Vector{ProteinChain{T}}: a collection of ProteinChains.

ProteinChains.ProteinStructureStore — Type

ProteinStructureStore <: AbstractDict{String,ProteinStructure}

A mutable struct representing a store for protein structures using HDF5 file format.

The struct implements the AbstractDict interface, allowing for dictionary-like operations.

A ProteinStructureStore gets closed when there no longer exists a program-accessible reference to it.

Examples

julia> store = ProteinStructureStore("store.h5")
ProteinStructureStore with 0 entries

julia> store["3HFM"] = pdb"3HFM"
[ Info: Downloading file from PDB: 3HFM
3-chain ProteinStructure{Float64} "3HFM.cif"
 215-residue ProteinChain{Float64, @NamedTuple{}} (H)
 214-residue ProteinChain{Float64, @NamedTuple{}} (L)
 129-residue ProteinChain{Float64, @NamedTuple{}} (Y)

julia> store
ProteinStructureStore with 1 entry

julia> keys(store)
Set{String} with 1 element:
  "3HFM"

julia> delete!(store, "3HFM")
ProteinStructureStore with 0 entries

ProteinChains.ProteinStructureStore — Type

ProteinStructureStore(filename, mode="cw")

Open or create an HDF5 file as a ProteinStructureStore where mode is one of:

"r" read only

"r+" read and write

"cw" read and write, create file if not existing, do not truncate

"w" read and write, create a new file (destroys any existing contents)

Structure attributes

The ProteinStructureStore type allows for storing metadata about each structure in the file, that can be lazily read from the file without loading the structure into memory using readattribute. These attributes are only written to the file when the structure is written, and are not preserved in memory after the structure has been read.

T: type of the coordinates in the structure.

n_residues: vector of number of residues per model in the structure.

n_chains: vector of number of chains per model in the structure.

ProteinChains.ProteinStructureStore — Method

ProteinStructureStore(f::Function, filename, mode="cw")

ProteinChains.append_residue — Method

append_residue(Backbone::Backbone, torsion_angles::Vector{<:Real}; ideal::BackboneGeometry=DEFAULT_BACKBONE_GEOMETRY)

Create a new backbone by appending 3 new torsion angles (ψ, ω, ϕ) at the end, using bond lengths and bond angles specified in BackboneGeometry.

ProteinChains.deleteproperty — Method

deleteproperty(store::ProteinStructureStore, name::AbstractString, index::Integer, property::Symbol)

Delete a property from a chain in a structure. Chains are indexed from 1 to length(store, name).

ProteinChains.deserialize — Method

deserialize(filename::AbstractString)

Deserialize ProteinStructure objects from an HDF5 file. Returns a Vector{ProteinStructure} of all structures stored in the file.

ProteinChains.mapmmcif — Method

mapmmcif(mmcifdict, field1 => field2, field3 => field4, ...)

julia> import BioStructures

julia> filename = BioStructures.downloadpdb("3HFM", format=BioStructures.MMCIFFormat);
[ Info: Downloading file from PDB: 3HFM

julia> mmcifdict = BioStructures.MMCIFDict(filename);

julia> mapmmcif(mmcifdict,
           "_atom_site.auth_asym_id"   => "_atom_site.label_entity_id",
           "_entity_src_gen.entity_id" => "_entity_src_gen.pdbx_gene_src_ncbi_taxonomy_id")
Dict{String, String} with 3 entries:
  "Y" => "9031"
  "L" => "10090"
  "H" => "10090"

ProteinChains.pdbentry — Method

pdbentry(pdbid::AbstractString; format=MMCIFFormat, kwargs...)

Keyword arguments get propagated to BioStructures.downloadpdb

Downloads are cached in a temporary directory.

Examples

julia> pdbentry("1EYE")
[ Info: Downloading file from PDB: 1EYE
1-chain ProteinStructure{Float64} "1EYE.cif"
 256-residue ProteinChain{Float64, @NamedTuple{}} (A)

julia> pdb"1EYE" # string macro for convenience
[ Info: File exists: 1EYE
1-chain ProteinStructure{Float64} "1EYE.cif"
 256-residue ProteinChain{Float64, @NamedTuple{}} (A)

julia> pdb"1EYE"A # string suffix to get a specific chain
[ Info: File exists: 1EYE
256-residue ProteinChain{Float64, @NamedTuple{}} (A)

julia> pdb"1EYE"1 # integer suffix to specify "ba_number" keyword
[ Info: Downloading file from PDB: 1EYE
2-chain ProteinStructure{Float64} "1EYE_ba1.cif"
 256-residue ProteinChain{Float64, @NamedTuple{}} (A)
 256-residue ProteinChain{Float64, @NamedTuple{}} (A-2)

ProteinChains.prepend_residue — Method

append_residue(Backbone::Backbone, torsion_angles::Vector{<:Real}; ideal::BackboneGeometry=DEFAULT_BACKBONE_GEOMETRY)

Create a new backbone by prepending 3 new torsion angles (ψ, ω, ϕ) at the beginning, using bond lengths and bond angles specified in the BackboneGeometry.

Note

The torsion angle order is the same as it would be when appending. The order is not reversed.

ProteinChains.readattribute — Method

readattribute(store::ProteinStructureStore, name::AbstractString, attribute::AbstractString)

Read metadata about a structure without loading the structure into memory.

ProteinChains.readproperty — Method

readproperty(store::ProteinStructureStore, name::AbstractString, index::Integer, property::Symbol)

Read a property from a chain in a structure. Chains are indexed from 1 to length(store, name).

ProteinChains.serialize — Method

serialize(filename::AbstractString, structures::AbstractVector{<:ProteinStructure})

Serialize a vector of ProteinStructure objects to an HDF5 file. This function creates a new ProteinStructureStore and writes each structure in the input vector to it. Each structure is stored using its name as the key.

ProteinChains.writeattribute — Method

writeattribute(store::ProteinStructureStore, name::AbstractString, attribute::AbstractString, value)

Write metadata about a structure.

ProteinChains.writeproperty — Method

writeproperty(store::ProteinStructureStore, name::AbstractString, index::Integer, property::Symbol, value)

Write a property to a chain in a structure. Chains are indexed from 1 to length(store, name).